# Conflicts:
#	src/main/resources/templates/card-list.html
#	src/main/resources/templates/card-search.html
#	src/main/resources/templates/login.html
#	src/main/resources/templates/register.html

+++ src/main/java/org/example/thedeckforge/entity/interfaces/ICardRepository.java
@@ -11,7 +11,7 @@ public interface ICardRepository {
     void populateCardList();
     List<Card> returnCardList();
     List<Card> returnCardListByName(String searchCriteria);
-    Optional<Card> returnCardById(int id);
+    Optional<Card> returnCardById(long id);
     Optional<Card> returnCardByName(String name);
     Optional<Card> returnCardByType(String type);
     Optional<Card> returnIllegalStateOfCard();
